policy/ai: use string value for JSON (#435)
Using a struct Value causes data conversion issues
for JSON numbers that are represented as float64
in protobuf that need to be deserialized to int/uint
types. Instead of implementing a custom deserializer
for the Request type, we treat the string as raw
JSON that can be deserialized directly to the Request.

Also fixes the serde_json::Value parsing which used
to_value instead of from_str previously(before 8d56f65),
which caused the defaults/overrides feature to never work since
its introduction. `from_str` is required to deserialize
the JSON string to its appropriate JSON value.

Signed-off-by: Shashank Ram <shashank.ram@solo.io>